refactor(functions): put each function in their own file
